gdbserver: Exit gdbserver on a kill packet when started with --stdin
When we run a gdbserver taking commands from stdin then this
will (almost) always be a slave to a specific gdb process.  When we
receive a kill packet this is gdb telling us that it is done with this
process, and we should exit.

However, when gdbserver is taking commands over a socket then we want to
remain alive after a kill so we can reconnect.

This commit adds a new parameter to the GdbServer object that controls
how it should respond to a kill packet.

ChangeLog:

	* server/GdbServer.cpp (GdbServer::GdbServer): Add new parameter,
	initialise member variable.
	(GdbServer::rspClientRequest): Modify handling of kill packet.
	* server/GdbServer.h (GdbServer::KillBehaviour): New enum.
	(GdbServer::GdbServer): Add new parameter.
	(GdbServer::killBehaviour): New member variable.
	* server/main.cpp (main): Pass suitable GdbServer::KillBehaviour
	value to new GdbServer object.