Reduce use of global optind
The global optind is part of the getopt framework, however, relying on
the value of this global to remain unchanged long after we've finished
parsing the command line arguments is a bad idea.  If we reenter the
getopt framework then the value of optind will be changed.

One reason we might have for reentering the getopt framework is if a new
target makes use of getopt (hint, a new target is coming that does just
this).

The solution is to take a local copy of optind once we've finished
parsing the command line arguments, and then use the local copy.  With
this done we're free to enter the getopt framework again.

ChangeLog:

	* server/main.cpp (main): Use new local variable instead of global
	optind where possible.