Rename timeout to mTimeout in gdbserver
There was some confusion when setting the timeout due to a local
'timeout' masking the instance variable 'timeout'.  Also, timeout is a
std::chrono::duration <double> rather than an int.  As a result, user
specified timeouts were basically ignored.

I rename the member variable with an 'm' prefix, which is inline with
most of the other member variables, and I've fixed the code for setting
the timeout.

ChangeLog:

	* server/GdbServerImpl.cpp (GdbServerImpl::GdbServerImpl): Update
	member variable name.
	(GdbServerImpl::rspContinue): Likewise.
	(GdbServerImpl::rspCommand): Likewise.
	* server/GdbServerImpl.h (GdbServerImpl::timeout): Renamed to...
	(GdbServerImpl::mTimeout): ...this.