Turn on -Wextra compile flag by default
Some of the targets force -Wextra and so it can be pretty annoying to
run into code that builds fine for the rest of the project, but suddenly
doesn't compile when included into a -Wextra using target.

So, I've tested 32 and 64 bit builds with -Wextra and on my 6.3.1
GCC everything builds fine, so this commit turns on -Wextra by default.

ChangeLog:

	* configure.ac: Turn -Wextra on by default.
	* configure: Regenerate.