Handle syscalls when single stepping
With this commit we should now be able to single step through syscalls.
After each syscall we need to know if we should finish off a step or a
continue (depends what we were doing when we hit the syscall), there's a
new variable to track this choice.

As we shouldn't get nested syscalls there's error checking in place to
ensure that the variable is in the default (no syscall in flight) state
when the target hits a syscall.

The one exception is the exit syscall, when the target exits we don't
get a syscall reply, and so the variable is left in the syscall is in
progress state.  The result is that when a gdb session connects, the
first syscall of the new session will trigger a warning about a syscall
already being in flight (the exit from the previous gdb session).  To
prevent this I specifically clear the syscall tracking variable for each
new connection.

ChangeLog:

	* server/GdbServerImpl.cpp (GdbServerImpl::GdbServerImpl):
	Initialise mSyscallContinuation.
	(GdbServerImpl::rspServer): Reset mSyscallContinuation after each
	new connection.
	(GdbServerImpl::rspSyscallRequest): Take parameter, update
	mSyscallContinuation.
	(GdbServerImpl::rspSyscallReply): Hangle mSyscallContinuation,
	don't return a bool.
	(GdbServerImpl::rspContinue): Pass parameter to rspSyscallRequest.
	(GdbServerImpl::rspSingleStep): Call rspSyscallRequest to handle
	syscalls.
	(GdbServerImpl::rspClientRequest): Return after processing syscall
	reply.
	* server/GdbServerImpl.h (GdbServerImpl::lastPacketType): Delete.
	(GdbServerImpl::SyscallContinuationType): New enum type.
	(GdbServerImpl::mSyscallContinuation): New member variable.
	(GdbServerImpl::rspSyscallRequest): Update parameter list.
	(GdbServerImpl::rspSyscallReply): Update return type.