gdbserver: Compiler with -faligned-new flag
Later compilers (at least in 7.2.1) require a new flag, -faligned-new
in order to support c++ new for types with large alignment
requirements.  This commit adds that flag to the CXXFLAGS.

In order to not break builds on older compilers that don't have
support for this flag, the new flag is only added to the CXXFLAGS if
it is supported by the compiler.

In order to make checking for compiler flags easier, some new m4
macros are added.

ChangeLog:

	* configure.ac: Add check for -faligned-new, and add to CXXFLAGS.
	* m4/cxx_flags_check.m4: New file.
	* Makefile.in: Regenerate.
	* aclocal.m4: Regenerate.
	* configure: Regenerate.
	* server/Makefile.in: Regenerate.
	* targets/Makefile.in: Regenerate.
	* targets/common/Makefile.in: Regenerate.
	* targets/gdbsim/Makefile.in: Regenerate.
	* targets/pegasus/Makefile.in: Regenerate.
	* targets/ri5cy/Makefile.in: Regenerate.
	* trace/Makefile.in: Regenerate.