gdbserver: Don't expect a reply to an exit syscall
The exit syscall is not part of GDB's remote syscall mechanism, and we
certainly don't expect to get an 'F' reply packet back as a result of
an exit.

For syscalls that are handled with the 'F' packet mechanism we store
whether we were stepping or continuing when we hit the syscall, and
this state is also used to detect nested syscall.  When we get the 'F'
reply from GDB we reset the stored state.

As the 'exit' syscall is not part of the 'F' syscall mechanism,
storing the stepping or continuing state leads to false warnings about
nested syscalls, when one core exits.

This commit fixes this issue by not storing the stepping or continuing
state for exit syscalls.

ChangeLog:

	* server/GdbServerImpl.cpp (GdbServerImpl::rspSyscallRequest):
	Don't store a continuation state for exit syscall.