When using stdin connection for gdbserver, detect stdin closing
When we use the StreamConnection class to manage a stdin/stdout based
rsp connection with gdb, we need to detect the case where gdb
disappears under our feet.

Most of the infra-structure is already in place to support this, the
only things we need to do is actually spot that stdin has closed, and
then remember that the out GdbServer driver code has called close on
the StreamConnection.  If we do this, and then refuse to reconnect the
StreamConnection, the outer GdbServer code will take care of
terminating for us.

ChangeLog:

	* server/StreamConnection.cpp
	(StreamConnection::StreamConnection): Initialise mIsConnected.
	(StreamConnection::rspClose): Update header comment, set
	mIsConnected.
	(StreamConnection::isConnected): Return mIsConnected.
	(StreamConnection::getRspCharRaw): Return -1 when zero bytes are
	read, this should only happen when stdin is closed.
	* server/StreamConnection.h (StreamConnection::mIsConnected):
	Declare new member variable.